requirements.txt
asgiref==3.9.2
certifi==2025.8.3
cffi==2.0.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.4.3
cryptography==46.0.1
daphne==4.1.0
Django==5.2.6
django-allauth==65.11.2
django-cors-headers==4.9.0
django-sslserver==0.22
djangorestframework==3.16.1
git-filter-repo==2.47.0
idna==3.10
oauthlib==3.3.1
pillow==11.3.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.2.1
requests==2.32.5
requests-oauthlib==2.0.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
